Tests/LibWeb/Text/expected/HTML/navigator-getBattery.txt

same promise: true
constructor: BatteryManager
charging: true
chargingTime: 0
dischargingTime: Infinity
level: 1
